backend/requirements.txt
flask
flask-cors
google-cloud-aiplatform
langchain
langchain-google-vertexai
gunicorn
pyjwt
cryptography
pyyaml